Rename base formatter file and update imports
